Fixed distance sensor

+++ Core/Inc/distance.h
@@ -14,8 +14,11 @@
 **************************************************************************************************/
 #ifndef _distance_h_
 #define _distance_h_
+
+#include "main.h"
+#include "helper.h"
 //Function to return the distance of the object 
 //No input 
 //returns float
-float distance() ;
+uint8_t distance() ;
 #endif /*distance_h_ */
